requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
chess==1.11.2
chex==0.1.89
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
distro==1.9.0
etils==1.12.2
filelock==3.18.0
flax==0.9.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
humanize==4.12.3
idna==3.10
importlib_resources==6.5.2
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mlproj_manager==0.0.31
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.2
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==2.7.1
opt_einsum==3.4.0
optax==0.2.5
orbax-checkpoint==0.11.16
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.2
ray==2.51.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.4
rich==14.0.0
rpds-py==0.28.0
scipy==1.16.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tensorstore==0.1.75
TextArena==0.7.4
toolz==1.0.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstable==0.5.1
urllib3==2.5.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
